Consolidate extension/utility functions that were duplicated into MDUtils
