+++ core/src/main/java/com/afollestad/materialdialogs/internal/main/DialogLayout.kt
@@ -22,7 +22,7 @@ import androidx.annotation.ColorInt
 import com.afollestad.materialdialogs.MaterialDialog
 import com.afollestad.materialdialogs.R
 import com.afollestad.materialdialogs.internal.button.DialogActionButtonLayout
-import com.afollestad.materialdialogs.utils.dimenPx
+import com.afollestad.materialdialogs.utils.MDUtil.dimenPx
 
 val DEBUG_COLOR_PINK = Color.parseColor("#EAA3CF")
 val DEBUG_COLOR_DARK_PINK = Color.parseColor("#E066B1")
