+++ files/src/main/java/com/afollestad/materialdialogs/files/utilext/ViewExt.kt
@@ -10,20 +10,3 @@ import android.view.View
 internal fun <T : View> T.setVisible(visible: Boolean) {
   visibility = if (visible) View.VISIBLE else View.INVISIBLE
 }
-
-internal fun <T : View> T.updatePadding(
-  left: Int = this.paddingLeft,
-  top: Int = this.paddingTop,
-  right: Int = this.paddingRight,
-  bottom: Int = this.paddingBottom
-) {
-  if (left == this.paddingLeft &&
-      top == this.paddingTop &&
-      right == this.paddingRight &&
-      bottom == this.paddingBottom
-  ) {
-    // no change needed, don't want to invalidate layout
-    return
-  }
-  this.setPadding(left, top, right, bottom)
-}
